Create HealthPickupSystem
This program asks the user if they are at full health; if they're not, it asks them if they would like to pick up the health pack. If the answer is "yes" or "Yes", the user will be fully healed. This was created as a practice test for creating UI.